Rather than timestamp, use a more sensible name
This new unique name is obtained by using the DateTimeOriginal of the jpeg file
as well as its timestamp. It is more sensible in the fact that it gives
information about date of the image being clicked to the user. Its human
readable. Better than some random numbers in the timestamp.
The format of this unique name is Year_Month_Date_Timestamp[last 6 digits].jpeg

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>